requirements.txt
aiogram==3.13.0
python-dotenv==1.0.0
aiohttp==3.9.1
pyngrok==7.1.8
fastapi==0.104.1
uvicorn==0.24.0
openai==1.12.0
